Move `StateListItem` to `Icinga\Module\Icingadb\Widget\ItemList`
