+++ library/Icingadb/Widget/ItemList/BaseServiceListItem.php
@@ -7,7 +7,6 @@ namespace Icinga\Module\Icingadb\Widget\ItemList;
 use Icinga\Module\Icingadb\Common\Links;
 use Icinga\Module\Icingadb\Common\NoSubjectLink;
 use Icinga\Module\Icingadb\Model\Service;
-use Icinga\Module\Icingadb\Widget\StateListItem;
 use ipl\Html\Attributes;
 use ipl\Html\Html;
 use ipl\Html\HtmlElement;
